Refactoring code, ready to create AddRepair screen
