+++ app/src/main/java/com/example/repairtool/login/LoginPage.kt
@@ -20,7 +20,6 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import androidx.core.content.ContextCompat.startActivity
 import com.example.repairtool.R
 import com.example.repairtool.volunteer.RepairActivity
