+++ app/src/main/java/com/example/repairtool/repairs/RepairList.kt
@@ -1,7 +1,7 @@
 package com.example.repairtool.repairs
 
 object RepairList {
-    var repairList = mutableListOf<Repair>(
+    var repairList = mutableListOf(
         Repair("Repair 1", "Main building", "Registered"),
         Repair("Repair 2", "Office","Registered"),
         Repair("Repair 3", "Toilets", "Pending")
