+++ app/src/main/java/com/example/repairtool/ui/theme/Theme.kt
@@ -34,7 +34,7 @@ private val LightColorPalette = lightColors(
 @Composable
 fun RepairToolTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
-    content: @Composable() () -> Unit
+    content: @Composable () -> Unit
 ) {
     val colors = if (darkTheme) {
         DarkColorPalette
